+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+## [Unreleased]
+
+### Fixed
+- Fixed the first cell's top border being clipped when the file is scrolled to the very top.
+
 ## [0.5.0] - 2026-05-22
 
 ### Added